chore(package): rename misnamed `servicestrades` package
Rename the incorrect package `servicestrades` to `services.trade`.
Update imports.
