+++ trading-app-demo/src/main/java/com/pcs/tradingapp/services/trade/TradeService.java
@@ -1,4 +1,4 @@
-package com.pcs.tradingapp.servicestrade;
+package com.pcs.tradingapp.services.trade;
 
 import java.util.List;
